refactor(redesign-plugin-menu): Use Plugins Context provider inside RowEditor instead of InnerDocument
