+++ packages/editor/src/core/inner-document.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useRef, useState } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
 
-import { PluginSelectionMenuContextProvider } from './contexts/plugins-context'
 import { EditorChildren } from './editor-children'
 import { useBlurOnOutsideClick } from './hooks/use-blur-on-outside-click'
 import type { EditorProps } from './types'
@@ -107,9 +106,7 @@ export function InnerDocument({ children, onChange, ...props }: EditorProps) {
 
   return (
     <div className="relative" ref={wrapperRef}>
-      <PluginSelectionMenuContextProvider>
-        <EditorChildren>{children}</EditorChildren>
-      </PluginSelectionMenuContextProvider>
+      <EditorChildren>{children}</EditorChildren>
     </div>
   )
 }
