+++ packages/editor/src/plugins/rows/components/add-row-button-large.tsx
@@ -1,18 +1,27 @@
+import { PluginSelectionMenuContext } from '@editor/core/contexts/plugins-context'
 import { faCirclePlus } from '@fortawesome/free-solid-svg-icons'
 import { FaIcon } from '@serlo/frontend/src/components/fa-icon'
 import { useEditorStrings } from '@serlo/frontend/src/contexts/logged-in-data-context'
+import { useContext } from 'react'
 
 interface AddRowButtonLargeProps {
-  onClick: React.MouseEventHandler<HTMLButtonElement>
+  insertPluginCallback: (pluginType: string, insertIndex?: number) => void
+  insertIndex: number
 }
 
-export function AddRowButtonLarge({ onClick }: AddRowButtonLargeProps) {
+export function AddRowButtonLarge({
+  insertIndex,
+  insertPluginCallback,
+}: AddRowButtonLargeProps) {
   const rowsStrings = useEditorStrings().plugins.rows
+  const pContext = useContext(PluginSelectionMenuContext)
 
   return (
     <button
       className="serlo-button-editor-secondary mx-auto mt-24 block"
-      onClick={onClick}
+      onClick={() => {
+        pContext.openSuggestions(insertPluginCallback, insertIndex)
+      }}
       data-qa="add-new-plugin-row-button"
     >
       <FaIcon icon={faCirclePlus} className="text-xl" />{' '}
