+++ packages/editor/src/plugins/rows/components/row-editor.tsx
@@ -1,7 +1,9 @@
+import { AddPluginModal } from '@editor/core/components/add-plugin-modal/add-plugin-modal'
+import { PluginSelectionMenuContext } from '@editor/core/contexts/plugins-context'
 import { StateTypeReturnType } from '@editor/plugin'
 import { editorPlugins } from '@editor/plugin/helpers/editor-plugins'
 import { selectIsFocused, useAppSelector } from '@editor/store'
-import { useRef } from 'react'
+import { useContext, useRef } from 'react'
 
 import { AddRowButtonFloating } from './add-row-button-floating'
 import type { RowsPluginConfig, RowsPluginState } from '..'
@@ -9,24 +11,25 @@ import { EditorRowRenderer } from '../editor-renderer'
 
 interface RowEditorProps {
   config: RowsPluginConfig
-  onAddButtonClick(index: number): void
   index: number
   rows: StateTypeReturnType<RowsPluginState>
   row: StateTypeReturnType<RowsPluginState>[0]
   hideAddButton?: boolean
+  insertPluginCallback: (pluginType: string, insertIndex?: number) => void
 }
 
 export function RowEditor({
   config,
-  onAddButtonClick,
   index,
   row,
   rows,
   hideAddButton = false,
+  insertPluginCallback,
 }: RowEditorProps) {
   const focused = useAppSelector((state) => selectIsFocused(state, row.id))
   const plugins = editorPlugins.getAllWithData()
   const dropContainer = useRef<HTMLDivElement>(null)
+  const pContext = useContext(PluginSelectionMenuContext)
 
   return (
     <div
@@ -35,6 +38,7 @@ export function RowEditor({
       // bigger drop zone with padding hack
       className="rows-child relative -ml-12 pl-12"
     >
+      <AddPluginModal />
       <EditorRowRenderer
         config={config}
         row={row}
@@ -47,7 +51,7 @@ export function RowEditor({
         focused={focused}
         onClick={(event: React.MouseEvent) => {
           event.preventDefault()
-          onAddButtonClick(index + 1)
+          pContext.openSuggestions(insertPluginCallback, index + 1)
         }}
         hide={hideAddButton}
       />
